feat: improve diff representation
